bug 1393832: Move other pytest tests
